Unify element_types for dim2 and dim3
